Move Transitions related classes to Animations directory [skip ci]
